Update index.md as per 12.13.12 team meeting
Added several additional fields to the core required / basic fields to support components of the open data policy (last update, contact name/email).  Added field "Homepage URL" in the list of expanded fields to allow for Data.gov to point users to an Agency-specific website for a dataset/API if the Agency has their own catalog or other website for the dataset.